wordcloud 0.2v/demo/myproject/requirements.txt

async-generator    ==1.10      
attrs              ==22.2.0    
beautifulsoup4     ==4.12.0
bs4                ==0.0.1
certifi            ==2022.6.15
cffi               ==1.15.1
charset-normalizer ==2.0.12
click              ==8.1.3
colorama           ==0.4.5
contourpy          ==1.0.7
cryptography       ==40.0.1
cycler             ==0.11.0
exceptiongroup     ==1.1.0
fake-useragent     ==1.1.3
Flask              ==2.2.3
Flask-SQLAlchemy   ==3.0.3
fonttools          ==4.38.0
greenlet           ==2.0.2
h11                ==0.14.0
idna               ==3.3
install            ==1.3.5
itsdangerous       ==2.1.2
Jinja2             ==3.1.2
joblib             ==1.2.0
JPype1             ==1.4.1
kiwisolver         ==1.4.4
konlpy             ==0.6.0
lxml               ==4.9.2
MarkupSafe         ==2.1.1
matplotlib         ==3.7.0
nltk               ==3.8.1
numpy              ==1.23.0
outcome            ==1.2.0
packaging          ==23.0
pandas             ==1.5.3
Pillow             ==9.1.1
pip                ==23.0.1
pycparser          ==2.21
PyMySQL            ==1.0.2
pyparsing          ==3.0.9
PySocks            ==1.7.1
python-dateutil    ==2.8.2
pytz               ==2022.7.1
regex              ==2022.10.31
requests           ==2.28.0
scikit-learn       ==1.2.1
scipy              ==1.10.1
selenium           ==4.8.2
setuptools         ==58.1.0
six                ==1.16.0
sniffio            ==1.3.0
sortedcontainers   ==2.4.0
soupsieve          ==2.4
split-folders      ==0.5.1
SQLAlchemy         ==2.0.4
threadpoolctl      ==3.1.0
torch              ==1.12.0
torchvision        ==0.13.0
tqdm               ==4.65.0
trio               ==0.22.0
trio-websocket     ==0.9.2
typing_extensions  ==4.2.0
urllib3            ==1.26.9
useragent          ==0.1.1
Werkzeug           ==2.2.3
wordcloud          ==1.8.1
wsproto            ==1.2.0